+++ build.sh
@@ -31,10 +31,10 @@ cat > "$CONTENTS/Info.plist" <<'PLIST'
 <dict>
     <key>CFBundleName</key>             <string>Cookie Monster</string>
     <key>CFBundleDisplayName</key>      <string>Cookie Monster</string>
-    <key>CFBundleIdentifier</key>       <string>com.cookiemonster.usage</string>
+    <key>CFBundleIdentifier</key>       <string>com.pflugpeil.cookiemonster</string>
     <key>CFBundleExecutable</key>       <string>CookieMonster</string>
-    <key>CFBundleVersion</key>          <string>0.2.1</string>
-    <key>CFBundleShortVersionString</key><string>0.2.1</string>
+    <key>CFBundleVersion</key>          <string>0.3.0</string>
+    <key>CFBundleShortVersionString</key><string>0.3.0</string>
     <key>CFBundlePackageType</key>      <string>APPL</string>
     <key>CFBundleIconFile</key>         <string>AppIcon</string>
     <key>LSMinimumSystemVersion</key>   <string>13.0</string>